requirements.txt
requests
python-dotenv
flask
gunicorn
gspread
pandas
openpyxl
anthropic
apscheduler
xhtml2pdf
